samples: wifi: Fix proto type used for RAW sockets
It turned out that IPPROTO_RAW is not a valid proto type for AF_PACKET
sockets, so it likely won't be supported anymore in the future.

As the samples doing raw TX aren't actually using IP protocols, but just
sending raw L2 frames, we could just use ETH_P_ALL as protocol when
creating socket instead.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>