applications: nrf5340_audio: fix headset_dfu configuration
The FILE_SUFFIX=release was intended to go into `ipc_radio` image.
The applications/nrf5340_audio/sysbuild/ipc_radio/prj_release.conf
should be used, but the file
applications/nrf5340_audio/sysbuild/ipc_radio/prj.conf
was used. This commit fixes that.

The FILE_SUFFIX=fota still is applied to the mcuboot image and the
main application image as it was before.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>