samples: bluetooth: add error message on overstay
In case someone modifies the nrf_dm sample, he might increase the
amount of time needed in the timeslot. If that is the case, the
customer would just see an non-understandable MPSL assertion error.
Improved this by printing an error message. There will still be an MPSL
assertion, but hopefully with this people will know how to resolve it.

Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>