nrf_rpc_uart: decode ACKs in ISR
Currently, decoding both normal packets and ACKs happens
in the receive thread. This imposes certain limitations,
for example the receive thread shall not send any packets
as this would introduce a chicken-egg problem: the thread
is waiting for an ACK but the ACK can only be decoded by
the current thread.

More importantly, this can also introduce spurious ACK
timeouts and retransissions because the receive thread may
be blocked when trying to push a packet to the thead pool.

Move the ACK decoding to ISR to overcome the mentioned
issues.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>