bluetooth: fast_pair: fmdn: add uptime workaround to the clock module
Added the kernel uptime workaround to the clock module in the FMDN
extension that is part of the Fast Pair library.

Applied the workaround for the kernel uptime persistence after a system
reset. The FMDN clock module assumes that the kernel uptime starts from
zero after the system reboot.

Activated the workaround only for devices from the nRF54L series that
use the GRTC timer as the system clock source (the default
configuration). For this configuration, the kernel uptime persists
after a system reset if the reset reason is of a specific type, such as
a software reset (for example, triggered by the sys_reboot API).

Ref: NCSDK-32268

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>