modules: memfault: Add support for uploading modem traces to Memfault
Add support for uploading modem traces to Memfault upon a coredump.
This is useful to get the full picture of what happened leading up to
a crash.

The feature is optional and guarded by a Kconfig option. It is enabled
by default in the overlay `overlay-modem_trace_to_memfault.conf`
added to the memfault sample.

Tests and docs has been updated accordingly.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>