lib: modem_slm: Stop triggering indicate multiple times
When indicate handler is registered and indicate pin configured,
thousands of callback calls were executed because the GPIO cannot
be triggering on an edge due to Hw issues increasing power consumption
Now there will be one callback call for one instance of
incoming data or AT notifications in SLM idle state.

This solution is not perfect but better than currently with
thousands of callbacks.

Jira: LRCS-82

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>