samples: bluetooth: Enable PSA RNG on 54H in the LLPM sample
The PSA RNG is a cryptographically secure random number generator.
It will be enabled by default, eventually.  For now, enable it manually.

This sample requires an increase in the number of buffers available
for incoming HCI events from the controller, a workaround for a
possible flow control issue in the HCI over IPC implementation.

Signed-off-by: Gordon Klaus <gordon.klaus@nordicsemi.no>