samples: cellular: location/modem_shell: Increase HEAP for Wi-Fi
PR #20438 changed kernel heap allocations because new Wi-Fi heap was
introduced (CONFIG_NRF_WIFI_CTRL_HEAP_SIZE).
Kernel heap was set to 10kB. This might have been for Wi-Fi positioning.
When combined "Cellular + Wi-Fi" positioning is done kernel heap
requirement for the data to send to cloud increases.
It seems 10kB heap is not enough and now it will be increased to 12kB.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>