lib: Implied NRF_FORCE_RAM_ON_REBOOT by RAM_POWER_DOWN_LIBRARY
Ram power down library can lead to the firmware crash after
reboot, if all RAM blocks are not powered before boot.

Application can enable NRF_FORCE_RAM_ON_REBOOT to solve this,
but it's better to take care of enabling it directly in the
library.

Thanks to that workaround can be removed from all Matter samples.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>