nrf_security: Fix for PSA MAC context in Cracen
There is an input_buffer in the Cracen MAC driver
which is used to store intermediate data when needed.
The size of this buffer was chosen based on the maximum
hash block size. Since there was some hash optimizations
merged earlier the maximum hash block size can now be 1
and this is problematic for cases where only CMAC is being
used.

This fixes this by setting the buffer accortingly when only
CMAC is enabled.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>