bluetooth: services: Implement real-time ranging data service
The RRSP now always supports the two operation modes: on-demand
and real-time data. When using real-time data, the CS procedure results
are notified immediately as soon as they are available. The RREQ doesn't
need to fetch or ACK the notifications, or wait for the data complete
indication.

In the case where new procedure results are received from the controller
before the previous procedure results have been completely sent, the new
results are dropped.

A callback has been exposed for the CCCD write request responses as it's
now possible for the RRSP to reject these if they would lead to both
on-demand and real-time modes to be enabled at the same time. The
alternative would be to block on a semaphore before returning from
bt_ras_rreq_realtime_rd_subscribe() or
bt_ras_rreq_on_demand_rd_subscribe().

Signed-off-by: Olivier Lesage <olivier.lesage@nordicsemi.no>