samples: tfm: nRF54L15 support for the secure peripheral sample
This adds support for the nRF54L15 with the secure peripheral
sample.

It refactors the secure peripheral sample and makes the
platform specific peripherals used part of the configuration
so that they can be used by all the platforms.

Along with that align the configuration to always use
CONFIG_SOC_SERIES... for both the code and CMake logic
and make sure that you fail to build when TFM is not enabled.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>