samples: bluetooth: Fix errata 117 check
There is a typo in the nrf5340 errata 117 logic, where it checked
for a nRF52-series device instead of a nRF53-series device.
This caused a hard-fault if the function ran successfully
on a nRF52-series device.

Fix is to use the same implementation as done in sample
../peripheral/radio_test

File and function:
radio_test.c::errata_117()

Signed-off-by: Håkon Alseth <haal@nordicsemi.no>