bluetooth: gatt_dm: silence compiler warning for unused parameter
When compiling with -Wunused-parameter the compiler emits a warning:

gatt_dm.h: In function 'bt_gatt_dm_data_print':
gatt_dm.h:329:67: warning: unused parameter 'dm' [-Wunused-parameter]

when bt_gatt_dm_data_print() is compiled but the corresponding Kconfig
CONFIG_BT_GATT_DM_DATA_PRINT=y is not set, which can be the case when
this Kconfig is only set for debug builds (and not release builds).

Since the parameter dm is intentionally unused when
CONFIG_BT_GATT_DM_DATA_PRINT is not defined it is okay to silence the
compiler warning.

Signed-off-by: Bryce Wilkins <bryce.wilkins@gmail.com>