net: lib: rest_client: Fix partial reply processing
The partial response processing in rest_client was broken, the current
callback logic would only work if the response was received in a single
chunk. In fact, we cannot share the REST reply buffer with the HTTP
client lib, as the latter will overwrite the buffer content after the
partial response callback call.

Fix this be defining an intermediate small buffer for the HTTP client
library for data reporting via registered callback function. The buffer
content is then copied to the large response buffer provided by the REST
lib, where the entire response is supposed to be collected. In case the
response data is larger than the buffer provided by the REST lib, the
response is truncated.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>