applications: nrf_desktop: Unsupported HID report IDs in HID reportq
Change updates HID reportq utility to accept HID report IDs which do not
belong to HID input reports supported by the application (are not part
of the `input_reports` array). This is done to improve robustness.

Change also aligns HID forward module to the updated API of the HID
reportq utility.

Jira: NCSDK-33592

Signed-off-by: Marek Pieta <Marek.Pieta@nordicsemi.no>
Signed-off-by: Divya Pillai <divya.pillai@nordicsemi.no>