modules: openthread: New OpenThread radio implementation for nRF
Implemented the new OpenThread radio implementation for the
nRF Connect platform.
The new radio implementation is based on the nRF IEEE802.15.4 radio
driver and does not use Zephyr's IEEE802.15.4 shim layer.
In this solution, the Zephyr's networking L2 layer is disabled, and
radio packets are passed directly to the nRF radio.

Prepared a build environment to support building OpenThread without
Zephyr's L2 and IEEE802.15.4 shim layers.

To use the new radio implementation, the CONFIG_NET_L2_OPENTHREAD
Kconfig must be set to `n`.

Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>
Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>