caf: leds: improve logic for calculating color count from DTS
Improved the CAF LEDs module by simplifying the logic for color
count calcuation that was done using the DTS configuration.
Now, it is no longer necessary to define a dummy array for
calculating the number of supported colors for each LED instance.
This improvement also resolves the build warnings for the LLVM
toolchain variant.

Ref: NCSDK-34195

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>