doc: changelog: fast_pair: fmdn: mention proper fix for FMDN clock issue
Added a changelog entry that describes removal of a workaround for the
issue where the Find My Device Network (FMDN) clock value is not set
correctly after the system reboot for nRF54L devices. The kernel uptime
value that is returned by the k_uptime_get function is now correctly set
to zero during the system bootup process for each reset type. As
a result, the workaround for the FMDN clock value is no longer needed.

Ref: NCSDK-33206

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>