samples: bluetooth: Fix nRF54H20 errata 216 workaround in DTM
Fix the workaround implementation in the direct_test_mode sample.
Fixed issues:
- RX mode: fix the assertion that happened after receiving a packet
- TX mode: fix scheduling of packet transmission that was affected
by the errata workaround and the number of packets was reduced.

The errata 216 'on' signal is sent upon reception of a command from
tester that starts radio (in RX or TX mode). The errata 216 'off'
signal is sent when a 'stop test' command is received.
This approach deviates from the recommended HMPAN-216 workaround
where the 'off' signal should be sent always when RADIO is disabled
for at least 100 us to avoid too big increase in the current
consumption. In DTM current consumption is not critical, so
sending errata 216 signals upon commands from the tester seems
better as RF test execution is not affected.

Ref. NCSDK-33915

Signed-off-by: Michał Grochala <michal.grochala@nordicsemi.no>