bluetooth: rpc: fix enable, disable, enable scenario
Fix the scenario in which a user calls bt_enable, followed
by bt_disable and bt_enable again. In such a case, BT RPC
client would not request loading Bluetooth settings for the
second bt_enable call, which is necessary to either load or
create Bluetooth identities and complete the initialization.

The faulty behavior was likely due to an attempt to optimize
registering static callbacks and services but:
1. Registering static callbacks and services was moved some
   time ago and it is now done before invoking bt_enable on
   the remote device, so the optimization isn't effective
   anymore.
2. Loading settings must happen for each bt_enable call as
   bt_disable clears all the identities.

Btw, also change the code for invoking settings load to
depend on BT_SETTINGS instead of more generic SETTINGS.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>