mpsl: Use nrf2 cloc ctrl resolve API to get LFCLK accuracy
When MPSL uses nrf2 clock control e.g. on nRF54h20, is should
call nrf_clock_control_resolve to get information about the
LFCLK source accuracy. The LFCLK souce is unknown to the MPSL
so it asks for the worst alowed LFCLK accuracy, which is 500PPM.
The API should provide the most accurate and the most power
efficient clock source.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>