bluetooth: hci_driver: Add missing mpsl.h include
Due to change in Zephyr hal, use of mpsl_hwres.h the mpsl.h
is not included silently with nrf header files. That causes
some functions to be not declared in the hci_driver.c.

Also the mpsl_pm_util.c requires additional header:
nrf_errno.h provided in nrfxlib with MPSL header files.
The header is required because integration layer uses
error codes from the headed in the sdk-nrf and MPSL
integration layer.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>