applications: nrf_desktop:  auto provision KMU targets
Updated the configurations of the nRF Desktop application for
board targets that store the MCUboot verification key in the KMU
peripheral. These nRF54L-based targets now use the automatic
provisioning feature that makes it possible to perform the KMU
provisioning together with the west flash operation. The provisioning
is only performed if the ``west flash`` command is executed with the
``--erase`` or ``--recover`` flag.

Aligned the sample documentation with this change.

JIRA: NCSDK-34240

Signed-off-by: Jan Zyczkowski <jan.zyczkowski@nordicsemi.no>